exploit.tools refactored into separate modules to avoid circular dependencies while using telemetries
