+++ monkey/infection_monkey/exploit/sshexec.py
@@ -6,11 +6,12 @@ import StringIO
 
 import infection_monkey.monkeyfs as monkeyfs
 from infection_monkey.exploit import HostExploiter
-from infection_monkey.exploit.tools import get_target_monkey, get_monkey_depth
+from infection_monkey.exploit.tools.helpers import get_target_monkey, get_monkey_depth, build_monkey_commandline
 from infection_monkey.model import MONKEY_ARG
 from infection_monkey.network.tools import check_tcp_port
-from infection_monkey.exploit.tools import build_monkey_commandline
 from common.utils.exploit_enum import ExploitType
+from common.utils.attack_utils import ScanStatus
+from infection_monkey.telemetry.attack.t1105_telem import T1105Telem
 
 __author__ = 'hoffer'
 
@@ -162,10 +163,11 @@ class SSHExploiter(HostExploiter):
                 ftp.putfo(file_obj, self._config.dropper_target_path_linux, file_size=monkeyfs.getsize(src_path),
                           callback=self.log_transfer)
                 ftp.chmod(self._config.dropper_target_path_linux, 0o777)
-
+                T1105Telem(ScanStatus.USED, self.host.ip_addr[0], src_path).send()
             ftp.close()
         except Exception as exc:
             LOG.debug("Error uploading file into victim %r: (%s)", self.host, exc)
+            T1105Telem(ScanStatus.SCANNED, self.host.ip_addr[0], src_path).send()
             return False
 
         try:
