+++ monkey/infection_monkey/exploit/tools/wmi_tools.py
@@ -1,0 +1,150 @@
+import logging
+
+from impacket.dcerpc.v5.dcom import wmi
+from impacket.dcerpc.v5.dcom.wmi import DCERPCSessionError
+from impacket.dcerpc.v5.dcomrt import DCOMConnection
+from impacket.dcerpc.v5.dtypes import NULL
+
+__author__ = 'itamar'
+
+LOG = logging.getLogger(__name__)
+
+
+class DceRpcException(Exception):
+    pass
+
+
+class AccessDeniedException(Exception):
+    def __init__(self, host, username, password, domain):
+        super(AccessDeniedException, self).__init__("Access is denied to %r with username %s\\%s and password %r" %
+                                                    (host, domain, username, password))
+
+
+class WmiTools(object):
+    class WmiConnection(object):
+        def __init__(self):
+            self._dcom = None
+            self._iWbemServices = None
+
+        @property
+        def connected(self):
+            return self._dcom is not None
+
+        def connect(self, host, username, password, domain=None, lmhash="", nthash=""):
+            if not domain:
+                domain = host.ip_addr
+
+            dcom = DCOMConnection(host.ip_addr,
+                                  username=username,
+                                  password=password,
+                                  domain=domain,
+                                  lmhash=lmhash,
+                                  nthash=nthash,
+                                  oxidResolver=True)
+
+            try:
+                iInterface = dcom.CoCreateInstanceEx(wmi.CLSID_WbemLevel1Login,
+                                                     wmi.IID_IWbemLevel1Login)
+            except Exception as exc:
+                dcom.disconnect()
+
+                if "rpc_s_access_denied" == exc.message:
+                    raise AccessDeniedException(host, username, password, domain)
+
+                raise
+
+            iWbemLevel1Login = wmi.IWbemLevel1Login(iInterface)
+
+            try:
+                self._iWbemServices = iWbemLevel1Login.NTLMLogin('//./root/cimv2', NULL, NULL)
+                self._dcom = dcom
+            except:
+                dcom.disconnect()
+
+                raise
+            finally:
+                iWbemLevel1Login.RemRelease()
+
+        def close(self):
+            assert self.connected, "WmiConnection isn't connected"
+
+            self._iWbemServices.RemRelease()
+            self._iWbemServices = None
+
+            self._dcom.disconnect()
+            self._dcom = None
+
+    @staticmethod
+    def dcom_wrap(func):
+        def _wrapper(*args, **kwarg):
+            try:
+                return func(*args, **kwarg)
+            finally:
+                WmiTools.dcom_cleanup()
+
+        return _wrapper
+
+    @staticmethod
+    def dcom_cleanup():
+        for port_map in DCOMConnection.PORTMAPS.keys():
+            del DCOMConnection.PORTMAPS[port_map]
+        for oid_set in DCOMConnection.OID_SET.keys():
+            del DCOMConnection.OID_SET[port_map]
+
+        DCOMConnection.OID_SET = {}
+        DCOMConnection.PORTMAPS = {}
+        if DCOMConnection.PINGTIMER:
+            DCOMConnection.PINGTIMER.cancel()
+            DCOMConnection.PINGTIMER.join()
+            DCOMConnection.PINGTIMER = None
+
+    @staticmethod
+    def get_object(wmi_connection, object_name):
+        assert isinstance(wmi_connection, WmiTools.WmiConnection)
+        assert wmi_connection.connected, "WmiConnection isn't connected"
+
+        return wmi_connection._iWbemServices.GetObject(object_name)[0]
+
+    @staticmethod
+    def list_object(wmi_connection, object_name, fields=None, where=None):
+        assert isinstance(wmi_connection, WmiTools.WmiConnection)
+        assert wmi_connection.connected, "WmiConnection isn't connected"
+
+        if fields:
+            fields_query = ",".join(fields)
+        else:
+            fields_query = "*"
+
+        wql_query = "SELECT %s FROM %s" % (fields_query, object_name)
+
+        if where:
+            wql_query += " WHERE %s" % (where,)
+
+        LOG.debug("Execution WQL query: %r", wql_query)
+
+        iEnumWbemClassObject = wmi_connection._iWbemServices.ExecQuery(wql_query)
+
+        query = []
+        try:
+            while True:
+                try:
+                    next_item = iEnumWbemClassObject.Next(0xffffffff, 1)[0]
+                    record = next_item.getProperties()
+
+                    if not fields:
+                        fields = record.keys()
+
+                    query_record = {}
+                    for key in fields:
+                        query_record[key] = record[key]['value']
+
+                    query.append(query_record)
+                except DCERPCSessionError as exc:
+                    if 1 == exc.error_code:
+                        break
+
+                    raise
+        finally:
+            iEnumWbemClassObject.RemRelease()
+
+        return query
