+++ monkey/infection_monkey/exploit/vsftpd.py
@@ -7,8 +7,8 @@
 import socket
 import time
 from infection_monkey.exploit import HostExploiter
-from infection_monkey.exploit.tools import build_monkey_commandline
-from infection_monkey.exploit.tools import get_target_monkey, HTTPTools, get_monkey_depth
+from infection_monkey.exploit.tools.helpers import get_target_monkey, build_monkey_commandline, get_monkey_depth
+from infection_monkey.exploit.tools.http_tools import HTTPTools
 from infection_monkey.model import MONKEY_ARG, CHMOD_MONKEY, RUN_MONKEY, WGET_HTTP_UPLOAD, DOWNLOAD_TIMEOUT
 from logging import getLogger
