+++ monkey/infection_monkey/exploit/weblogic.py
@@ -8,7 +8,8 @@
 from __future__ import print_function
 from requests import post, exceptions
 from infection_monkey.exploit.web_rce import WebRCE
-from infection_monkey.exploit.tools import get_free_tcp_port, get_interface_to_target
+from infection_monkey.exploit.tools.helpers import get_interface_to_target
+from infection_monkey.network.info import get_free_tcp_port
 from BaseHTTPServer import BaseHTTPRequestHandler, HTTPServer
 
 import threading
