+++ monkey/infection_monkey/monkey.py
@@ -19,7 +19,7 @@ from infection_monkey.telemetry.attack.victim_host_telem import VictimHostTelem
 from infection_monkey.windows_upgrader import WindowsUpgrader
 from infection_monkey.post_breach.post_breach_handler import PostBreach
 from common.utils.attack_utils import ScanStatus
-from infection_monkey.exploit.tools import get_interface_to_target
+from infection_monkey.exploit.tools.helpers import get_interface_to_target
 
 __author__ = 'itamar'
