+++ monkey/infection_monkey/post_breach/actions/users_custom_pba.py
@@ -6,6 +6,9 @@ from infection_monkey.post_breach.pba import PBA
 from infection_monkey.control import ControlClient
 from infection_monkey.config import WormConfiguration
 from infection_monkey.utils import get_monkey_dir_path
+from infection_monkey.telemetry.attack.t1105_telem import T1105Telem
+from common.utils.attack_utils import ScanStatus
+from infection_monkey.exploit.tools.helpers import get_interface_to_target
 
 LOG = logging.getLogger(__name__)
 
@@ -81,7 +84,13 @@ class UsersPBA(PBA):
 
         if not pba_file_contents or not pba_file_contents.content:
             LOG.error("Island didn't respond with post breach file.")
+            T1105Telem(ScanStatus.SCANNED,
+                       get_interface_to_target(WormConfiguration.current_server.split(':')[0]),
+                       filename).send()
             return False
+        T1105Telem(ScanStatus.USED,
+                   get_interface_to_target(WormConfiguration.current_server.split(':')[0]),
+                   filename).send()
         try:
             with open(os.path.join(dst_dir, filename), 'wb') as written_PBA_file:
                 written_PBA_file.write(pba_file_contents.content)
