+++ monkey/infection_monkey/windows_upgrader.py
@@ -8,7 +8,7 @@ import time
 import infection_monkey.monkeyfs as monkeyfs
 from infection_monkey.config import WormConfiguration
 from infection_monkey.control import ControlClient
-from infection_monkey.exploit.tools import build_monkey_commandline_explicitly
+from infection_monkey.exploit.tools.helpers import build_monkey_commandline_explicitly
 from infection_monkey.model import MONKEY_CMDLINE_WINDOWS
 from infection_monkey.utils import is_windows_os, is_64bit_windows_os, is_64bit_python
